Fix all warnings, and better separate code that should not be mixed
Clean up a number of warnings in the tree.

Refactor especially the core code so we don't end up building a bunch
of BIOS-specific code for EFI, and general build cleanups.

+++ com32/elflink/ldlinux/chainboot.c
@@ -61,7 +61,7 @@ void chainboot_file(const char *file, uint32_t type)
     reg.eax.l = max;
     reg.ebx.l = 0;
     reg.edx.w[0] = 0;
-    reg.edi.l = (uint32_t)buf;
+    reg.edi.l = (uintptr_t)buf;
     reg.ebp.l = -1;	/* XXX: limit? */
     reg.esi.w[0] = rv;
